[PROD-325] Extract extraction code, remove support for parsed log
