+++ tests/test_eventlog.py
@@ -1,5 +1,7 @@
 import json
 import tempfile
+import unittest
+import zipfile
 from pathlib import Path
 
 from spark_log_parser import eventlog
@@ -15,8 +17,6 @@ def test_simple_emr_log():
 
         with open(log.event_log) as log_fobj:
             event = json.loads(log_fobj.readline())
-
-    assert not log.is_parsed, "Log is correctly identified as being not parsed"
 
     assert all(key in event for key in ["Event", "Spark Version"]), "All keys are present"
 
@@ -34,84 +34,10 @@ def test_simple_databricks_log():
         with open(log.event_log) as log_fobj:
             event = json.loads(log_fobj.readline())
 
-    assert not log.is_parsed, "Log is correctly identified as being not parsed"
-
     assert all(
         key in event
         for key in ["Event", "Spark Version", "Timestamp", "Rollover Number", "SparkContext Id"]
     ), "All keys are present"
-
-    assert event["Event"] == "DBCEventLoggingListenerMetadata", "Expected first event is present"
-
-
-def test_databricks_rollover_log():
-    event_log = Path("tests", "logs", "databricks-rollover.zip").resolve()
-
-    with tempfile.TemporaryDirectory(
-        prefix="eventlog-%s-" % event_log.name[: -len("".join(event_log.suffixes))]
-    ) as temp_dir:
-        log = eventlog.EventLog(source_url=event_log.as_uri(), work_dir=temp_dir)
-
-        with open(log.event_log) as log_fobj:
-            event = json.loads(log_fobj.readline())
-            log_fobj.seek(0)
-
-            rollover_count = 0
-            for i, event_str in enumerate(log_fobj):
-                event = json.loads(event_str)
-                if "Rollover Number" in event:
-                    assert all(
-                        key in event
-                        for key in [
-                            "Event",
-                            "Spark Version",
-                            "Timestamp",
-                            "Rollover Number",
-                            "SparkContext Id",
-                        ]
-                    ), "All keys are present"
-                    assert (
-                        rollover_count == event["Rollover Number"]
-                    ), "Contiguous monotonically increasing IDs"
-                    rollover_count += 1
-
-            assert rollover_count == 3, "All log parts are present"
-            assert i + 1 == 16945, "All events are present"
-
-
-def test_databricks_messy_rollover_log_archive():
-    event_log = Path("tests", "logs", "databricks-rollover-messy.zip").resolve()
-
-    with tempfile.TemporaryDirectory(
-        prefix="eventlog-%s-" % event_log.name[: -len("".join(event_log.suffixes))]
-    ) as temp_dir:
-        log = eventlog.EventLog(source_url=event_log.as_uri(), work_dir=temp_dir)
-
-        with open(log.event_log) as log_fobj:
-            event = json.loads(log_fobj.readline())
-            log_fobj.seek(0)
-
-            rollover_count = 0
-            for i, event_str in enumerate(log_fobj):
-                event = json.loads(event_str)
-                if "Rollover Number" in event:
-                    assert all(
-                        key in event
-                        for key in [
-                            "Event",
-                            "Spark Version",
-                            "Timestamp",
-                            "Rollover Number",
-                            "SparkContext Id",
-                        ]
-                    ), "All keys are present"
-                    assert (
-                        rollover_count == event["Rollover Number"]
-                    ), "Contiguous monotonically increasing IDs"
-                    rollover_count += 1
-
-            assert rollover_count == 3, "All log parts are present"
-            assert i + 1 == 16945, "All events are present"
 
 
 def test_raw_databricks_log():
@@ -131,21 +57,51 @@ def test_raw_databricks_log():
     ), "All keys are present"
 
     assert event["Event"] == "DBCEventLoggingListenerMetadata", "Expected first event is present"
+    assert event["Event"] == "DBCEventLoggingListenerMetadata", "Expected first event is present"
 
 
-def test_parsed_log():
-    parsed_path = Path("tests", "logs", "databricks-parsed.json").resolve()
+class RolloverLog(unittest.TestCase):
+    def test_databricks_rollover_log(self):
+        self.validate_log(Path("tests", "logs", "databricks-rollover.zip").resolve(), 3, 16945)
 
-    with tempfile.TemporaryDirectory(
-        prefix="eventlog-%s-" % parsed_path.name[: -len("".join(parsed_path.suffixes))]
-    ) as temp_dir:
-        log = eventlog.EventLog(source_url=parsed_path.as_uri(), work_dir=temp_dir)
+    def test_databricks_messy_rollover_log_archive(self):
+        self.validate_log(
+            Path("tests", "logs", "databricks-rollover-messy.zip").resolve(), 3, 16945
+        )
 
-        with open(log.event_log) as log_fobj:
-            parsed = json.loads(log_fobj.readline())
+    def test_databricks_messy_rollover_log_dir(self):
+        with tempfile.TemporaryDirectory() as temp_dir:
+            with zipfile.ZipFile(Path("tests", "logs", "databricks-rollover-messy.zip")) as zinfo:
+                temp_dir_path = Path(temp_dir)
+                zinfo.extractall(temp_dir_path)
+                self.validate_log(temp_dir_path, 3, 16945)
 
-    assert log.is_parsed, "Parsed flag is set"
+    def validate_log(self, event_log: Path, log_file_total: int, log_entry_total: int):
+        with tempfile.TemporaryDirectory() as temp_dir:
+            log = eventlog.EventLog(source_url=event_log.as_uri(), work_dir=temp_dir)
 
-    assert all(
-        key in parsed for key in ["jobData", "stageData", "taskData"]
-    ), "All keys are present"
+            with open(log.event_log) as log_fobj:
+                event = json.loads(log_fobj.readline())
+                log_fobj.seek(0)
+
+                rollover_count = 0
+                for i, event_str in enumerate(log_fobj):
+                    event = json.loads(event_str)
+                    if "Rollover Number" in event:
+                        assert all(
+                            key in event
+                            for key in [
+                                "Event",
+                                "Spark Version",
+                                "Timestamp",
+                                "Rollover Number",
+                                "SparkContext Id",
+                            ]
+                        ), "All keys are present"
+                        assert (
+                            rollover_count == event["Rollover Number"]
+                        ), "Contiguous monotonically increasing IDs"
+                        rollover_count += 1
+
+                assert rollover_count == log_file_total, "All log parts are present"
+                assert i + 1 == log_entry_total, "All events are present"
